% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shylock-package.R
\docType{package}
\name{shylock-package}
\alias{shylock}
\alias{shylock-package}
\title{shylock: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
